feat: add mobile view of team list page
