+++ app/(general)/performances/[id]/teams/_components/TeamListTable/DeleteButton.tsx
