+++ components/Header/index.tsx
@@ -3,9 +3,9 @@ import Image from "next/image"
 import Link from "next/link"
 
 import Sidebar from "@/components/Header/_component/Sidebar"
+import ROUTES from "@/constants/routes"
+import { cn } from "@/lib/utils"
 
-import ROUTES from "../../constants/routes"
-import { cn } from "../../lib/utils"
 import NavLink from "../NavLink"
 import Profile from "./_component/Profile"
 
@@ -19,7 +19,7 @@ const Header = ({
   const menuItems: { name: string; url: string; active: boolean }[] = [
     { name: "공지사항", url: ROUTES.NOTICE.LIST.url, active: false },
     { name: "공연목록", url: ROUTES.PERFORMANCE.LIST.url, active: true },
-    { name: "세션지원", url: ROUTES.TEAM.LIST.url, active: true },
+    { name: "세션지원", url: ROUTES.PERFORMANCE.TEAMS(1).url, active: true },
     { name: "맴버목록", url: ROUTES.MEMBER.LIST.url, active: true }
   ]
