+++ components/StatusBadge.tsx
@@ -1,24 +1,26 @@
 import { Badge } from "@/components/ui/badge"
 import { cn } from "@/lib/utils"
 
-const StatusBadge = ({ status }: { status: "Inactive" | "Active" }) => {
-  const className =
-    status === "Inactive"
-      ? "bg-red-100 text-destructive"
-      : "bg-green-100 text-green-600"
+interface StatusBadgeProps {
+  status: "Inactive" | "Active"
+  className?: string
+}
+
+const StatusBadge = ({ status, className }: StatusBadgeProps) => {
   return (
-    <div className="flex justify-center">
-      <Badge
-        variant="outline"
-        className={cn(
-          className,
-          "text-md font-lig rounded-full border-none px-4 py-1"
-        )}
-      >
-        <span className="me-2 font-extrabold">●</span>
-        {status}
-      </Badge>
-    </div>
+    <Badge
+      variant="outline"
+      className={cn(
+        status === "Inactive"
+          ? "bg-red-100 text-destructive"
+          : "bg-green-100 text-green-600",
+        "text-md font-lig rounded-full border-none px-4 py-1",
+        className
+      )}
+    >
+      <span className="me-2 font-extrabold">●</span>
+      {status}
+    </Badge>
   )
 }
